requirements.txt
aniso8601==4.0.1
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
gevent==1.3.7
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jwcrypto==0.6.0
MarkupSafe==1.1.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.2
pytz==2018.7
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.15
urllib3==1.24.1
Werkzeug==0.14.1
